pyproject.toml
[tool.poetry]
name = "b-thesis"
version = "0.1.0"
description = "Bachelorthesis Pierre Bathoum"
authors = ["Pierre Bathoum <pierre.bathoum@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.4"
pandas = "^2.2.2"
matplotlib = "^3.8.4"
torch = "^2.3.0"
torchvision = "^0.18.0"
torchaudio = "^2.3.0"
scikit-learn = "^1.4.2"
seaborn = "^0.13.2"
notebook = "^7.1.3"
opencv-python = "^4.9.0.80"
ruptures = "^1.1.9"
h5py = "^3.11.0"
keras = "^3.4.0"
tensorflow = "^2.16.1"
pillow = "^10.3.0"
pyyaml = "^6.0.1"
easydict = "^1.13"
termcolor = "^2.4.0"
tqdm = "^4.66.4"
gradio-imageslider = "^0.0.20"
hdbscan = "^0.8.37"
lightning = "^2.4.0"
apex = "^0.9.10.dev0"
lightning-bolts = "^0.7.0"
lightly = "^1.5.12"
ffmpeg-python = "^0.2.0"
pywavelets = "^1.7.0"
plotly = "^5.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
